requirements.txt
# 数据获取
akshare>=1.14.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

# AI 分析
google-generativeai>=0.8.0
openai>=1.0.0

# 通知
python-telegram-bot>=20.0

# 调度
schedule>=1.2.0

# 工具
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
